# ARROW-10015: [Rust] Simd aggregate kernels