@jorgecarleitao 2020-10-07
This is really exciting!!

Just to check my understanding, we are talking `9.48 us` vs `250 ns` when we use this branch with SIMD vs current master without SIMD, which corresponds to a 36x overall improvement. Did you ran the current master with SIMD (to compare them with the same feature gate)?

I will start by reviewing the other branch first.

fyi @andygrove @alamb , this may have "some" impact on the query benchmarks of DataFusion.