@jhorstmann 2020-10-08
Rebased after merge of [ARROW-10040][1] (#8262)

 [1]: https://issues.apache.org/jira/browse/ARROW-10040