@andygrove 2020-10-08
I'm planning on running TPC-H benchmarks later today with and without this patch.